do not lock up addproc on worker setup timeout
This is a corollary to the previous commit in #32290, and implements 
suggestions thereof.

It restricts the master to wait for a worker to respond within 
`Distributed.worker_timeout()` seconds. Beyond that it releases the lock 
on `rr_ntfy_join` with a special flag `:TIMEDOUT`. This flag is set to 
`:ERROR` in case of any errors during worker setup, and to `:OK` when 
the master received a `JoinCompleteMsg` indicating setup completion from 
worker.

`addprocs` returns the worker id in the list of workers it added only if 
it has received a `JoinCompleteMsg`, that is, only when `rr_ntfy_join` 
contains `:OK`. Note that the worker process may not be dead yet, and it 
may still be listed in `workers()` until it actually goes down.